fix: deliver and draw the EMP charge ring
The invemp handler expected six tokens but the server sends five, so the charge ring never started. The payload now travels as integers because %f through this VM's variadic formatting truncated the message mid-float. The ring renders as a circle of world sprites (EMP_RING_STEPS in cg_local.h) instead of one billboard, which sat below the camera frustum and was culled. The cgame suite now checks the ring count, spread, growth and expiry (94 checks).

+++ code/game/g_invoke.c
@@ -348,9 +348,11 @@ static void G_InvokeCastSpell( gentity_t *ent, invokeState_t *st, int hand,
 	case SPELL_EMP:
 		VectorCopy( ps->origin, st->empOrigin );
 		st->empTime = level.time + 2500;
-		// the client draws the charge ring from this until the burst
-		trap_SendServerCommand( ent - g_entities, va( "invemp %f %f %f %i\n",
-			st->empOrigin[0], st->empOrigin[1], st->empOrigin[2], 2500 ) );
+		// the client draws the charge ring from this until the burst.
+		// Coordinates go as integers: float varargs are not safe through
+		// this VM's print/format path.
+		trap_SendServerCommand( ent - g_entities, va( "invemp %i %i %i %i\n",
+			(int)st->empOrigin[0], (int)st->empOrigin[1], (int)st->empOrigin[2], 2500 ) );
 		break;
 	case SPELL_CHAOS_METEOR:
 		VectorCopy( ps->origin, start );
